Replay Subject with additional period of time factor
Now you can choose to buffer a certain number of events.
Typically used for caching values and then replaying them for later
subscriptions.
Accepts a timewindow in milliseconds for which the events will be cached
after which newer events will be cached for the same duration.
During which any subscriber would get all the cached events.