Async Subject
tt is similar to the Replay and Behavior subjects in the way that it
caches values, however it will only store the last value,
and only publish it when the sequence is completed.
The general usage of the AsyncSubject<T> is to only ever publish one
 value then immediately complete. Almost like a Task.
So even if values:1, 2, 3 were emitted before calling complete(), the
subscriber only receives 3 after the call is complete.